fixes
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rgraphviz@3120 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/Rgraphviz.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-SEXP R_scalarReal(float v) {
+SEXP R_scalarReal(double v) {
     SEXP ans = allocVector(REALSXP,1);
     REAL(ans)[0] = v;
     return(ans);
